+++ sklearn/utils/__init__.py
@@ -8,6 +8,8 @@ import warnings
 from .validation import *
 from .murmurhash import murmurhash3_32
 
+# Make sure that DeprecationWarning get printed
+warnings.simplefilter("always", DeprecationWarning)
 
 class deprecated(object):
     """Decorator to mark a function or class as deprecated.